pyproject.toml
[project]
name = "anvesh"
version = "0.1.0"
description = "A powerful 'Blue Ocean' lead scraper to find high-value businesses with zero online presence."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "playwright>=1.57.0",
    "psycopg[binary]>=3.3.2",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "python-dotenv>=1.2.1",
    "uvicorn>=0.40.0",
]
